+++ master/char.h
@@ -11,7 +11,7 @@ void sendChar(uint8_t x,uint8_t y,uint8_t lx,uint8_t ly, const uint8_t Bstabe);
 
 void print(const char *text,uint8_t x,uint8_t y);
 void printN(const uint16_t *number,uint8_t x,uint8_t y);
-void printB(const uint8_t *number,uint8_t x,uint8_t y);
+void printB(const uint16_t *number,uint8_t x,uint8_t y);
 
 
 #endif
